refactor: move logger to web3 provider
