+++ packages/mask/src/extension/popups/UI.tsx
@@ -15,7 +15,7 @@ import { LoadingPlaceholder } from './components/LoadingPlaceholder/index.js'
 import Services from '../service.js'
 import { Web3ContextProvider } from '@masknet/web3-hooks-base'
 import { ProviderType } from '@masknet/web3-shared-evm'
-import { LoggerContextProvider, LogPlatform } from '@masknet/shared'
+import { LoggerContextProvider, LogPlatform } from '@masknet/web3-providers'
 
 function usePopupTheme() {
     return usePopupFullPageTheme(useValueRef(languageSettings))
