+++ packages/shared-base/src/index.ts
@@ -21,4 +21,3 @@ export * from './serializer/index.js'
 export * from './test-helpers/index.js'
 export * from './unsafe-html/index.js'
 export * from './utils/index.js'
-export * from './logger/type.js'
