+++ packages/shared/src/index.ts
@@ -7,4 +7,3 @@ export * from './locales/languages.js'
 export * from './constants.js'
 export * from './storage/index.js'
 export * from './types.js'
-export * from './sentry/index.js'
