+++ packages/web3-providers/src/Sentry/component.tsx
