+++ packages/web3-providers/src/Sentry/index.ts
